feat: add basic language definitions
Add enums for Keywords, Symbols and classes that define what a Number or
an Identifier are.

Those classes can also check if an array of chars conform to the
definition of whatever they represent (e.g. a Keyword), so they can be
used by the lexer to parse the source code.