feat: add basic primitive lexer
Add a very basic and primitive lexer.

It is hand-written, which means that it won't be super optimized or
fast, but it is way easier to understand.

It also doesn't use any DFA but instead just looks at arrays of chars
and checks if they are a valid keyword, symbol, where delimiters are and
so on, based on the check() functions in the language definitions.