fix: assignment operator backwards
Fix an error in the lexer, that recognized the assignment operator
backwards.

The assignment operator is represented as ":=", a colon, followed by an
equality sign. The lexer recognized it the other way around, "=:".

Fix the order in which the lexer tries to detect the chars that make up
the assignment operator.