feat: add basic (unfinished) ll(1) parser
Add a very basic LL(1) parser that produces an abstract syntax tree.

Since the grammar for pl/0 is neither left-recursive nor ambiguous, it
is possible to parse the language left-to-right and with a lookahead of
a single token, which are the properties that need to be fulfilled for
an LL(1) parser.

The SyntaxTree is built from many SyntaxTreeNodes, that each have a
single Terminal- or Nonterminal-Symbol attached to them.

While the parser walks through its different methods, it adds single
nodes to the tree.

The terminal and non-terminal symbols are defined as enums, currently
without a direct mapping onto the different tokens, that mapping also
occurs in the SyntaxTreeNode instances.

Currently, the parser lacks the ability to parse conditions or
while-loops, this will be added in a later commit.