fix: parser doesn't recognize end
Flip the while loop in which statements are parsed from a do-while to a
conventional while loop.

The do-while tries to parse another statement after the last one, since
it checks if there is a semicolon at the end, and if there is one, tries
to parse the next statement.

Using a simple while loop instead alleviates that problem.