feat: add basic code generator
The CodeGenerator traverses the AST depth-first, left-to-right and
generates the corresponding assembly instructions for the stack machine.

It also handles stuff like substituting constants with their assigned
values which could be considered an optimisation and thus may be done in
its one phase before the actual code generation in the future.