feat!: add two-char operators
To make the operators for "less or equal than", "not equal" etc. more
like the defacto standard in other programming languages, introduce the
ability to recognise and parse two-character long operators.

This requires a change in the definition of the language and thus must
be considered a breaking change.

BREAKING CHANGES:
- change operator definitions for "less or equal than", "greater or
  equal than", "not equal" to two-character sequences instead of the
  single characters