fix(parser): rework parser to correctly parse expressions
Change the different operations from Terminal to NonTerminal nodes, so
that the tree for expressions is built in a more cohesive and above all
correct way, even for more complex expressions and expressions involving
brackets.

The parser now expects all expressions in the token stream to be in
postfix notation instead of the infix notation that is used in the
program source code. This makes parsing them and creating the
corresponding syntax tree a way easier task, since operator precedence
isn't an issue anymore.

Refs: #PC-6