fix(generator): correctly generate code from the tree
Since the syntax tree now has the operators as non terminal nodes, the
code generation needs to be refactored to work correctly with the
changed syntax tree.

The changes made to the syntax tree actually make the code generation
somewhat simpler, so this is all in all more of a simplification.

Refs: #PC-6 Fixed